Add podium shot commands Co-authored-by: naveed1117 <naveed1117@users.noreply.github.com> Co-authored-by: jopy-man <jopy-man@users.noreply.github.com> Co-authored-by: Ardian Agoes <ardianagoes@users.noreply.github.com> Co-authored-by: BrianZ60 <BrianZ60@users.noreply.github.com> Co-authored-by: Lapcas <Lapcas@users.noreply.github.com> Co-authored-by: Zixi52 <Zixi52@users.noreply.github.com> Co-authored-by: reyamiller <reyamiller@users.noreply.github.com>

+++ src/main/java/com/stuypulse/robot/commands/shooter/HorizontalShooter/HorizontalShooterSetRPM.java
@@ -17,7 +17,7 @@ public class HorizontalShooterSetRPM extends InstantCommand {
 
     @Override
     public void initialize() {
-        horizontalShooter.setTargetRPM(topTargetRPM, bottomTargetRPM);
+        horizontalShooter.setTargetRPM(topTargetRPM.doubleValue(), bottomTargetRPM.doubleValue());
     }
